+++ Water_Features/ChangeLog.md
@@ -1,6 +1,8 @@
-﻿## Patch V1.1.11
+﻿## Update V1.2.0
 * Added a setting to attempt to force water simulation in game to be at least 1x.
 * Seasonal streams is by default off.
 * Seasonal Streams and Waves and Tides settings are only available in game or in the editor. All aspects of those settings are saved into save games now. Valid water sources are required or it will default to disabled.
 * Reorganized settings including making Water Tool tab as initial tab.
 * Additional settings to make waves and tides / seasonal streams work in the editor have been removed.
+* New! Automated Water Source: Custom modded water source that dynamically changes water source type to automatically fill and stabilize to the desired water surface elevation. Not affected by Seasonal Streams or Waves and Tides settings.
+* Map extents determined dynamically for compatibility with other mods.